tests/requirements.txt
https://github.com/druids/pynamodb/tarball/AddTagsSupport#egg=pynamodb
django-germanium==2.2.3
flake8
coverage
